pyproject.toml
[project]
name = "memolite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-chroma>=1.1.0",
    "langchain-siliconflow>=1.0.0",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
